Refactor Login component to its own directory
